# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/tokens": {
            "get": {
                "description": "Returns a paginated list of tokens discovered by the indexer. Supports filtering by name, symbol and ERC type.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "List tokens",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by token name (case-insensitive, partial match)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by token symbol (case-insensitive, partial match)",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "RAYLS_ERC20",
                            "RAYLS_ERC721",
                            "RAYLS_ERC1155",
                            "RAYLS_ENYGMA",
                            "RAYLS_ERC721_DVP",
                            "RAYLS_ERC1155_DVP",
                            "RAYLS_STABLECOIN"
                        ],
                        "type": "string",
                        "description": "Filter by ERC standard (canonical labels)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by the chain the token was deployed on (chainId)",
                        "name": "issuerId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api_adapters_handlers.tokenListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Deploys a token via RNContractFactory, signed with the authenticated user's custody wallet.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "Deploy a protocol token",
                "parameters": [
                    {
                        "description": "Token to deploy",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api_adapters_handlers.deployTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/cmd_api_adapters_handlers.deployTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ops-api_cmd_api_utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/ops-api_cmd_api_utils.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/tokens/estimate": {
            "post": {
                "description": "Returns the real gas estimate (eth_estimateGas + gas price) for deploying a token via RNContractFactory, signed by the authenticated user's custody wallet. Does not deploy.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "Estimate token deploy gas",
                "parameters": [
                    {
                        "description": "Token to estimate",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api_adapters_handlers.deployTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ops-api_cmd_api_core.TokenDeployEstimate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ops-api_cmd_api_utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cmd_api_adapters_handlers.deployTokenRequest": {
            "type": "object",
            "required": [
                "standard"
            ],
            "properties": {
                "decimals": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "standard": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
        "cmd_api_adapters_handlers.deployTokenResponse": {
            "type": "object",
            "properties": {
                "deployedAddress": {
                    "type": "string"
                },
                "standard": {
                    "type": "string"
                },
                "txHash": {
                    "type": "string"
                }
            }
        },
...
        "cmd_api_adapters_handlers.tokenListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cmd_api_adapters_handlers.tokenResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "cmd_api_adapters_handlers.tokenResponse": {
            "type": "object",
            "properties": {
                "contractAddress": {
                    "type": "string"
                },
                "decimals": {
                    "type": "integer"
                },
                "ercStandard": {
                    "type": "integer"
                },
                "ercStandardLabel": {
                    "type": "string"
                },
                "holderCount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "issuerId": {
                    "type": "string"
                },
                "metadataUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "statusLabel": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "tokenClass": {
                    "type": "string"
                },
                "totalSupply": {
                    "type": "string"
                }
            }
        },
...
        "ops-api_cmd_api_core.TokenDeployEstimate": {
            "type": "object",
            "properties": {
                "gasLimit": {
                    "type": "integer"
                },
                "gasPriceWei": {
                    "type": "string"
                },
                "totalFeeWei": {
                    "description": "gasLimit * gasPriceWei",
                    "type": "string"
                }
            }
        },
...
        "ops-api_cmd_api_utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "hint": {
                    "type": "string"
                }
            }
        }
    }
}